--- a/benchmarks_test.go
+++ b/benchmarks_test.go
@@ -287,6 +287,36 @@ func Benchmark_Cmp(bench *testing.B) {
 	bench.Run("single/uint256", benchmark_Cmp_Bit)
 }
 
+func BenchmarkLt(b *testing.B) {
+	benchmarkUint256 := func(b *testing.B, samples *[numSamples]Int) (flag bool) {
+		var x Int
+		for j := 0; j < b.N; j += numSamples {
+			for i := 0; i < numSamples; i++ {
+				y := samples[i]
+				flag = x.Lt(&y)
+				x = y
+			}
+		}
+		return
+	}
+	benchmarkBig := func(b *testing.B, samples *[numSamples]big.Int) (flag bool) {
+		var x big.Int
+		for j := 0; j < b.N; j += numSamples {
+			for i := 0; i < numSamples; i++ {
+				y := samples[i]
+				flag = x.Cmp(&y) < 0
+				x = y
+			}
+		}
+		return
+	}
+
+	b.Run("large/uint256", func(b *testing.B) { benchmarkUint256(b, &int256Samples) })
+	b.Run("large/big", func(b *testing.B) { benchmarkBig(b, &big256Samples) })
+	b.Run("small/uint256", func(b *testing.B) { benchmarkUint256(b, &int64Samples) })
+	b.Run("small/big", func(b *testing.B) { benchmarkBig(b, &big64Samples) })
+}
+
 func benchmark_Lsh_Big(n uint, bench *testing.B) {
 	original := big.NewInt(0).SetBytes(hex2Bytes("FBCDEF090807060504030201ffffffffFBCDEF090807060504030201ffffffff"))
 	bench.ResetTimer()