--- a/uint256_test.go
+++ b/uint256_test.go
@@ -17,6 +17,8 @@ import (
 var (
 	bigtt256 = new(big.Int).Lsh(big.NewInt(1), 256)
 	bigtt255 = new(big.Int).Lsh(big.NewInt(1), 255)
+
+	_ fmt.Formatter = &Int{} // Test if Int supports Formatter interface.
 )
 
 func hex2Bytes(str string) []byte {