```
name                      old time/op  new time/op  delta
_Add/single/uint256-6     3.08ns ± 0%  1.82ns ± 0%  -40.91%  (p=0.008 n=5+5)
_Sub/single/uint256-6     3.09ns ± 0%  1.82ns ± 0%  -41.10%  (p=0.008 n=5+5)
_Sub/single/uint256_of-6  1.82ns ± 0%  1.82ns ± 0%     ~     (all equal)
_Mul/single/uint256-6     15.9ns ± 0%  13.6ns ± 0%     ~     (p=0.079 n=4+5)
_Square/single/uint256-6  16.0ns ± 0%  14.6ns ± 0%   -8.27%  (p=0.008 n=5+5)
_And/single/uint256-6     1.67ns ± 5%  1.67ns ± 5%     ~     (p=1.000 n=5+5)
_Or/single/uint256-6      1.62ns ± 1%  1.61ns ± 1%     ~     (p=0.524 n=5+5)
_Xor/single/uint256-6     1.61ns ± 0%  1.66ns ± 5%     ~     (p=0.556 n=4+5)
_Cmp/single/uint256-6     2.96ns ± 0%  2.58ns ± 0%     ~     (p=0.079 n=4+5)
Lt/large/uint256-6        5.29ns ± 0%  2.53ns ± 0%  -52.21%  (p=0.008 n=5+5)
Lt/small/uint256-6        3.70ns ± 0%  2.53ns ± 0%  -31.48%  (p=0.008 n=5+5)
_Lsh/n_eq_0/uint256-6     4.09ns ± 0%  4.09ns ± 0%     ~     (all equal)
_Lsh/n_gt_192/uint256-6   4.32ns ± 0%  4.32ns ± 0%     ~     (all equal)
_Lsh/n_gt_128/uint256-6   5.50ns ± 0%  5.54ns ± 2%     ~     (p=0.079 n=4+5)
_Lsh/n_gt_64/uint256-6    7.30ns ± 0%  7.31ns ± 0%     ~     (p=0.722 n=5+5)
_Lsh/n_gt_0/uint256-6     8.19ns ± 0%  8.19ns ± 0%     ~     (all equal)
_Rsh/n_eq_0/uint256-6     4.09ns ± 0%  4.09ns ± 0%     ~     (all equal)
_Rsh/n_gt_192/uint256-6   4.32ns ± 0%  4.32ns ± 0%     ~     (all equal)
_Rsh/n_gt_128/uint256-6   5.56ns ± 0%  5.56ns ± 0%     ~     (p=0.333 n=4+5)
_Rsh/n_gt_64/uint256-6    7.23ns ± 0%  7.23ns ± 0%     ~     (all equal)
_Rsh/n_gt_0/uint256-6     8.19ns ± 0%  8.19ns ± 0%     ~     (all equal)
_Exp/large/uint256-6      6.89µs ± 0%  6.00µs ± 0%  -12.90%  (p=0.008 n=5+5)
_Exp/small/uint256-6       594ns ± 0%   516ns ± 0%  -13.13%  (p=0.008 n=5+5)
Div/small/uint256-6       13.0ns ± 0%  11.4ns ± 0%     ~     (p=0.079 n=4+5)
Div/mod64/uint256-6       57.4ns ± 0%  57.4ns ± 0%     ~     (p=0.444 n=5+5)
Div/mod128/uint256-6      89.3ns ± 0%  89.3ns ± 0%     ~     (p=0.397 n=5+5)
Div/mod192/uint256-6      80.1ns ± 0%  80.1ns ± 0%     ~     (p=0.429 n=5+4)
Div/mod256/uint256-6      67.6ns ± 0%  67.3ns ± 0%   -0.38%  (p=0.024 n=5+5)
Mod/small/uint256-6       14.8ns ± 0%  12.8ns ± 0%  -13.51%  (p=0.000 n=5+4)
Mod/mod64/uint256-6       59.1ns ± 0%  59.4ns ± 0%   +0.51%  (p=0.016 n=5+4)
Mod/mod128/uint256-6      91.0ns ± 0%  91.5ns ± 0%   +0.55%  (p=0.016 n=5+4)
Mod/mod192/uint256-6      81.7ns ± 0%  82.3ns ± 0%   +0.73%  (p=0.008 n=5+5)
Mod/mod256/uint256-6      69.7ns ± 0%  69.8ns ± 0%     ~     (p=0.167 n=5+5)
AddMod/small/uint256-6    18.7ns ± 2%  16.7ns ± 0%  -10.60%  (p=0.008 n=5+5)
AddMod/mod64/uint256-6     137ns ± 0%    67ns ± 0%  -51.17%  (p=0.000 n=5+4)
AddMod/mod128/uint256-6    212ns ± 0%   107ns ± 1%  -49.34%  (p=0.008 n=5+5)
AddMod/mod192/uint256-6    197ns ± 0%   100ns ± 0%  -49.24%  (p=0.008 n=5+5)
AddMod/mod256/uint256-6    157ns ± 0%    88ns ± 0%  -43.89%  (p=0.029 n=4+4)
MulMod/small/uint256-6    41.0ns ± 0%  37.4ns ± 0%   -8.78%  (p=0.008 n=5+5)
MulMod/mod64/uint256-6     101ns ± 0%   102ns ± 0%   +0.99%  (p=0.008 n=5+5)
MulMod/mod128/uint256-6    174ns ± 0%   176ns ± 0%   +1.15%  (p=0.016 n=4+5)
MulMod/mod192/uint256-6    172ns ± 0%   177ns ± 0%   +2.67%  (p=0.016 n=4+5)
MulMod/mod256/uint256-6    168ns ± 0%   168ns ± 0%     ~     (all equal)
_SDiv/large/uint256-6     90.3ns ± 0%  88.9ns ± 0%   -1.53%  (p=0.000 n=5+4)
```

Still don't understand why e.g. `_Mul/single/uint256-6     15.9ns ± 0%  13.6ns ± 0%     ~     (p=0.079 n=4+5)` is reported as "not affected".